separate lisa and custom optims
